Add a search session and log request params
We want to be able to associate all choices made for a single "session".
I'm not ready to implement everything necessary to authenticate users
and let them log in -- we'll want to do that eventually, but for now I
just want to be able to associate all of the choices made during the
same session of clicks to each other so that we can see for a given
person with a given appetite, what were all of the things that they
chose over the things that they didn't choose (and what were the things
that they didn't choose).