Add CORS header and redesign build for v0.2
In order to simplify the integration and deployment process, we wanted
to separate the static content (web pages and javascript) from the
dynamic backend (this dropwizard app). In the original release we got
around having to deal with Cross Origin Resource Sharing by running both
the webserver and the hunger web service on the same server AND having
the webserver act as a proxy to the web service for any external
requests. This worked, but hosting our static web pages from an ec2
micro instance seemed to be a little bit slower than letting github do
it for free with github pages. It also meant that deployment of changes
to the web pages was more involved than simply pushing those changes to
the master branch of the github repo.

Now that we've decoupled the two sets of resources, we can also simplify
the integration and deployment step which means we can bring the docker
build back into this repository rather than having it managed from a
separate integration repo. This may change again in the future, but it
works for now.