Implement a basic configuration for supporting stratification
For the next model, we want to start being more strategic about the
food pairings that we present to users. We gain little information when
we present a popular item next to an unpopular one, so we'd like to
start ranking images using click through rate as a basis for popularity.
Only foods in the same quantile (closely matching percentiles) should be
paired up together, so this will allow us to stratify by quantile.
Further, we'd also like to pair up items that already seem to be
disimilar, so each strata will also be divided along some abstract
dimension of similarity (using collaborative filtering). So items on the
top side of some abstract dimension of similarty will be least like
those on the bottom side in terms of some abstract characteristic.