Initial implementation of first model-b
I'm not particularly proud of this implementation, but I spent the
better part of a day or two trying to setup a whole abstract model
configuration when what we really want to do is pull that out into a
separate step anyway. A lot of details that should be configurable are
hardcoded at the moment, but I want to get something up and running
quickly so that we can go back and make it better later.