Initial commit for complete refactor
I'm trying to convert the applcation over to a serverless javascript
function that can run as a Google Firebase Cloud Function. The only file
that I'm maintaining for now is the original config.yml as it includes
the list of food items to pull from. We'll scrap the model-b
implementation for now and go back to making sure that we can at the
very least provide the original random based model with sufficient
logging of requests.