Add a stratified model/strategy
Split the sessions 50/50 between the original random model (one that
shows two random food images) and a random stratified model (one that
limits pairings to the same quantile based on click historical click
through rate). This may help improve the quality of clicks in the future
as we will be less likely to pair up popular items against unpopular
items. We'll need to recalculate quantiles and redeploy on a fairly
regular basis, especially if we add in new food items without much
history.