Refactor for angular feedme site
As we move over to an angular single page app, most of the logic for
deciding which foods to display has moved over to the client side
making the hunger service mostly obsolete. After going serverless, we
did notice the extra latency on the user side -- images took longer to
load because the hunger service took longer to load. That was pretty
easy to fix by pushing all of the logic over into the client
side/browser, but that meant that we had essentially architected
ourselves out of the opportunity to collect click data.

The appetite service is replacing the old hunger service as an
asyncronous call that simply logs anonymous appetite data voluntarily
sent by users -- as in they have explicitly been asked if they're ok
with sending it and they only send it for the purpose of us being able
to gather data to give everybody a better experience. It happens in the
background, so response time isn't particularly important, and now the
client side is able to batch up clicks before sending them, so we don't
need any kind of session id to match them up after the fact.